Crate iris_datatypes

Re-exports

pub use connection::ConnRecord;

pub use dns_transaction::DnsTransaction;

pub use http_transaction::HttpTransaction;

pub use quic_stream::QuicStream;

pub use ssh_handshake::SshHandshake;

pub use tls_handshake::TlsHandshake;

pub use conn_fts::*;

pub use maybe_quic::*;

pub use maybe_zoom::*;

pub use packet_list::*;

pub use static_type::*;

Modules

conn_fts

Various individual connection-level subscribable types for TCP and/or UDP connection information, statistics, and state history.

connection

A sample connection record that provides various TCP and/or UDP connection information, statistics, and state history. It does not deliver payload data.

dns_transaction

A DNS transaction. Subscribable alias for iris_core::protocols::stream::dns::Dns

http_transaction

An Http transaction. Subscribable alias for iris_core::protocols::stream::http::Http

maybe_quic

A streaming filter that heuristically accepts connections that look like QUIC: on an expected port, not clearly another protocol, and initial packets with (potential) QUIC short headers.

maybe_zoom

A streaming filter that heuristically accepts connections that look like Zoom: on a UDP port in ZOOM_PORT_RANGE, with initial packets whose first byte matches one of Zoom’s observed payload-type markers.

packet_list

Vectors of raw packet data. All of these types are Connection-level, meaning they are delivered on connection termination.

quic_stream

A Quic stream. Subscribable alias for iris_core::protocols::stream::quic::QuicConn

ssh_handshake

A SSH handshake. Subscribable alias for iris_core::protocols::stream::ssh::Ssh

static_type

Static-level datatypes. A data type is considered “static” if it can be inferred at or before the first packet in a connection and it stays constant throughout a connection. See datatypes, including foreign types, that implement StaticData.

tls_handshake

A TLS handshake. Subscribable alias for iris_core::protocols::stream::tls::Tls

Traits

FromSession

Need to define traits in this crate to avoid “cannot define inherent impl for foreign type” and “only traits defined in the current crate can be implemented for types defined outside of the crate” errors. Convenience method to convert a Session into a datatype that can be subscribed to. Datatypes implementing this trait are automatically Level=L7EndHdrs.

StaticData

Trait implemented by datatypes that are constant throughout a connection and inferrable at first packet (level is L4FirstPacket).