iris_datatypes

Trait FromSession

pub trait FromSession {
    // Required method
    fn from_session(session: &Session) -> Option<&Self>;
}

Need to define traits in this crate to avoid “cannot define inherent impl for foreign type” and “only traits defined in the current crate can be implemented for types defined outside of the crate” errors. Convenience method to convert a Session into a datatype that can be subscribed to. Datatypes implementing this trait are automatically Level=L7EndHdrs.

Required Methods

fn from_session(session: &Session) -> Option<&Self>

Build Self from a parsed session, or return None if impossible. Invoked when the session is fully matched, parsed, and ready to be delivered to a callback.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".